+++ pos_multi_session/static/src/js/pos_multi_session.js
@@ -8,11 +8,10 @@ odoo.define('pos_multi_session', function(require){
     var models = require('point_of_sale.models');
     var bus = require('bus.bus');
     var chrome = require('point_of_sale.chrome');
+    var longpolling = require('pos_longpolling');
 
     var _t = core._t;
 
-    // prevent bus to be started by chat_manager.js
-    bus.bus.activated = true; // fake value to ignore start_polling call
 
     screens.OrderWidget.include({
         rerender_orderline: function(order_line){
@@ -52,24 +51,17 @@ odoo.define('pos_multi_session', function(require){
             this.get('orders').bind('remove', function(order, collection, options){
                 if (!self.multi_session.client_online) {
                     if (order.order_on_server ) {
-                        var warning_message = _t("No connection to the server. You can only create new orders. It is forbidden to modify existing orders.");
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
                         self.multi_session.warning(warning_message);
                         return false;
                     }
                 }
                 order.ms_remove_order();
             });
-            this.ready.then(function () {
-                self.init_multi_session();
-            });
-
-        },
-        init_multi_session: function(){
-            if (this.config.multi_session_id){
-                this.multi_session = new exports.MultiSession(this);
-                this.multi_session.start();
-                this.multi_session.request_sync_all();
-            }
+            this.multi_session = new exports.MultiSession(this);
+            var channel_name = "pos.multi_session";
+            var callback = this.ms_on_update;
+            this.add_channel(channel_name, callback, this);
         },
         ms_my_info: function(){
             var user = this.cashier || this.user;
@@ -295,12 +287,12 @@ odoo.define('pos_multi_session', function(require){
             OrderSuper.prototype.initialize.apply(this, arguments);
             this.ms_info = {};
             this.revision_ID = options.revision_ID || 1;
+
             if (!_.isEmpty(options.ms_info)){
                 this.ms_info = options.ms_info;
             } else if (this.pos.multi_session){
                 this.ms_info.created = this.pos.ms_my_info();
             }
-
             this.ms_replace_empty_order = is_first_order;
             is_first_order = false;
             this.bind('change:sync', function(){
@@ -339,6 +331,8 @@ odoo.define('pos_multi_session', function(require){
                 this.pos.pos_session.order_ID = this.pos.pos_session.order_ID + 1;
                 this.sequence_number = this.pos.pos_session.order_ID;
                 this.trigger('change:update_new_order');
+            } else {
+                this.trigger('change');
             }
             if (!this.ms_check())
                 return;
@@ -349,7 +343,7 @@ odoo.define('pos_multi_session', function(require){
                 function(){
                     self.ms_update_timeout = false;
                     self.do_ms_update();
-                }, 1000);
+                }, 0);
         },
         ms_remove_order: function(){
             if (!this.ms_check())
@@ -374,33 +368,31 @@ odoo.define('pos_multi_session', function(require){
         },
         do_ms_update: function(){
             var self = this;
-            var data = this.export_as_JSON();
-            this.just_printed = false;
-            this.pos.multi_session.update(data).fail(function(error){
-                if (error == 'offline') {
-                    if (self.order_on_server) {
-                        self.order_on_server = true;
-                    } else {
-                        self.order_on_server = false;
+            if (this.enquied)
+                return;
+            var f = function(){
+                self.enquied=false;
+                var data = self.export_as_JSON();
+                return self.pos.multi_session.update(data).done(function(res){
+                    self.order_on_server = true;
+                    if (res) {
+                        var server_revision_ID = res.revision_ID;
+                        var order_ID = res.order_ID;
+                        if (self.sequence_number != order_ID) {
+                            self.sequence_number = order_ID;
+                            // sequence number replace
+                            self.pos.pos_session.order_ID = order_ID;
+                            // rerender order
+                            self.trigger('change');
+                        }
+                        if (server_revision_ID && server_revision_ID > self.revision_ID) {
+                            self.revision_ID = server_revision_ID;
+                        }
                     }
-                }
-            }).done(function(res){
-                self.order_on_server = true;
-                if (res) {
-                    var server_revision_ID = res.revision_ID;
-                    var order_ID = res.order_ID;
-                    if (self.sequence_number != order_ID) {
-                        self.sequence_number = order_ID;
-                        // sequence number replace
-                        self.pos.pos_session.order_ID = order_ID;
-                        // rerender order
-                        self.trigger('change');
-                    }
-                    if (server_revision_ID && server_revision_ID > self.revision_ID) {
-                        self.revision_ID = server_revision_ID;
-                    }
-                }
-            });
+                })
+            };
+            this.enquied = true;
+            this.pos.multi_session.enque(f);
         }
     });
     var OrderlineSuper = models.Orderline;
@@ -445,24 +437,31 @@ odoo.define('pos_multi_session', function(require){
 
     exports.MultiSession = Backbone.Model.extend({
         initialize: function(pos){
+            var self = this;
             this.pos = pos;
-            this.show_warning_message = true;
             this.client_online = true;
             this.order_ID = null;
-        },
-        start: function(){
-            var self = this;
-
-            this.bus = bus.bus;
-            this.bus.last = this.pos.db.load('bus_last', 0);
-            this.bus.on("notification", this, this.on_notification);
-            this.bus.stop_polling();
-            this.bus.start_polling();
-
+            this.update_queue = $.when();
+            this.func_queue = [];
+            this.pos.longpolling_connection.on("change:poll_connection", function(status){
+                if (status) {
+                    if (self.offline_sync_all_timer) {
+                        clearInterval(self.offline_sync_all_timer);
+                        self.offline_sync_all_timer = false;
+                    }
+                    self.request_sync_all();
+                } else {
+                    if (!self.offline_sync_all_timer) {
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
+                        self.warning(warning_message);
+                        self.start_offline_sync_timer();
+                    }
+                }
+            });
         },
         request_sync_all: function(){
             var data = {};
-            this.send({'action': 'sync_all', data: data});
+            return this.send({'action': 'sync_all', data: data});
         },
         remove_order: function(data){
             this.send({action: 'remove_order', data: data});
@@ -470,12 +469,28 @@ odoo.define('pos_multi_session', function(require){
         update: function(data){
             return this.send({action: 'update_order', data: data});
         },
+        enque: function(func){
+            var self = this;
+            this.func_queue.push(func);
+            this.update_queue = this.update_queue.then(function() {
+                if (self.func_queue[0]) {
+                    var next = $.Deferred();
+                    var func1 = self.func_queue.shift();
+                    func1().always(function () {
+                        next.resolve()
+                    });
+                    return next;
+                }
+            })
+        },
+        _debug_send_number: 0,
         send: function(message){
+            var current_send_number = 0;
             if (this.pos.debug){
-                console.log('MS', this.pos.config.name, 'send:', JSON.stringify(message));
+                current_send_number = this._debug_send_number++;
+                console.log('MS', this.pos.config.name, 'send #' + current_send_number +' :', JSON.stringify(message));
             }
             var self = this;
-            var connection_status = new $.Deferred();
             message.data.pos_id = this.pos.config.id;
             var send_it = function () {
                 return openerp.session.rpc("/pos_multi_session/update", {
@@ -483,27 +498,29 @@ odoo.define('pos_multi_session', function(require){
                     message: message
                 });
             };
-            send_it().fail(function (error, e) {
+            return send_it().fail(function (error, e) {
+                if (self.pos.debug){
+                    console.log('MS', self.pos.config.name, 'failed request #'+current_send_number+':', error.message);
+                }
                 if(error.message === 'XmlHttpRequestError ') {
                     self.client_online = false;
                     e.preventDefault();
-                    connection_status.reject('offline');
-                    if (self.show_warning_message) {
-                        var warning_message = _t("No connection to the server. You can only create new orders. It is forbidden to modify existing orders.");
+                    self.pos.longpolling_connection.network_is_off();
+                    if (!self.offline_sync_all_timer) {
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
                         self.warning(warning_message);
                         self.start_offline_sync_timer();
-                        self.show_warning_message = false;
                     }
                 } else {
                     self.request_sync_all();
                 }
             }).done(function(res){
+                if (self.pos.debug){
+                    console.log('MS', self.pos.config.name, 'response #'+current_send_number+':', JSON.stringify(res));
+                }
                 var server_orders_uid = [];
                 self.client_online = true;
-                if (res.action == "update_revision_ID") {
-                    connection_status.resolve(res);
-                }
-                connection_status.resolve();
+
                 if (res.action == "revision_error") {
                     var warning_message = _t('There is a conflict during synchronization, try your action again');
                     self.warning(warning_message);
@@ -515,16 +532,17 @@ odoo.define('pos_multi_session', function(require){
                         server_orders_uid.push(item.data.uid);
                     });
                     self.pos.pos_session.order_ID = res.order_ID;
-                    self.pos.pos_session.sequence_number = res.order_ID;
+
+                    if (res.order_ID != 0) {
+                        self.pos.pos_session.sequence_number = res.order_ID;
+                    }
                     self.destroy_removed_orders(server_orders_uid);
                 }
                 if (self.offline_sync_all_timer) {
                     clearInterval(self.offline_sync_all_timer);
                     self.offline_sync_all_timer = false;
                 }
-                self.show_warning_message = true;
             });
-            return connection_status;
         },
         destroy_removed_orders: function(server_orders_uid) {
             var self = this;
@@ -557,7 +575,7 @@ odoo.define('pos_multi_session', function(require){
             var self = this;
             var orders = this.pos.get("orders");
             orders.each(function(item) {
-                if (item.order_on_server === false) {
+                if (!item.order_on_server && item.get_orderlines().length > 0) {
                     item.ms_update();
                 }
             });
@@ -567,30 +585,6 @@ odoo.define('pos_multi_session', function(require){
             self.offline_sync_all_timer = setInterval(function(){
                 self.request_sync_all();
             }, 5000);
-        },
-        on_notification: function(notification) {
-            var self = this;
-            if (typeof notification[0][0] === 'string') {
-                notification = [notification];
-            }
-            for (var i = 0; i < notification.length; i++) {
-                var channel = notification[i][0];
-                var message = notification[i][1];
-                this.on_notification_do(channel, message);
-            }
-            this.pos.db.save('bus_last', this.bus.last);
-        },
-        on_notification_do: function (channel, message) {
-            if(Array.isArray(channel) && channel[1] === 'pos.multi_session'){
-                try{
-                    this.pos.ms_on_update(message);
-                }catch(err){
-                    this.pos.chrome.gui.show_popup('error',{
-                        'title': _t('Error'),
-                        'body': err,
-                    });
-                }
-            }
         }
     });
     return exports;