+++ pos_longpolling/doc/changelog.rst
@@ -1,3 +1,13 @@
+`1.1.1`
+-------
+
+- **FIX:** longpolling channel
+
+`1.1.0`
+-------
+
+- **NEW:** longpolling connection status in POS interface
+
 `1.0.0`
 -------