+++ pos_multi_session/doc/changelog.rst
@@ -15,6 +15,23 @@ Updates
 -------
 - FIX: For pos restaurant compatibility. Sync printed positions.
 
+`3.0.2`
+-------
+
+- FIX: Remove unpaid orders, once all synced sessions are closed.
+
+`3.0.1`
+-------
+
+- FIX: Issue on sessions synchronization after order transfer.
+
+`3.0.0`
+-------
+
+- FIX: Added a queue for request sending that allows to fix the syncronization error on slow or lost  connection
+- NEW: Added connection status with server to POS interface
+- NEW: Create new orders even if the connection with server temporarily has been lost
+
 `2.0.1`
 -------